+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 - Waveform color palette is derived from semantic token colors to better align with color theme
 - Added time units to save files
+- Viewer will show missing signals from save files not found in netlist
 - Clicking on an area outside the waveforms de-selects signals
+- Left/Right arrow no longer moves the marker if rows with no value change data are selected
 - Organized user settings into categories
 - Added user settings for default waveform colors
 - Added Epoch time (ns) number format
+- Fixed status bar not updating when setting alt marker, or selecting the last value change
+- Update wellen to 0.20.3
 
 ## 1.5.1 - 3/9/2026 - Latest Release